Added some changes of entities.

+++ src/main/java/allpccollector/repository/DomainUsersRepository.java
@@ -3,7 +3,7 @@ package allpccollector.repository;
 import allpccollector.model.DomainUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DomainUsersRepository extends JpaRepository<DomainUser, Integer>{
+public interface DomainUsersRepository extends JpaRepository<DomainUser, Long>{
 
     public DomainUser findDomainUserByUsername(String username);
 }
